refactor: list 컴포넌트 구조
- 각 부분별 함수 분리
