--- packages/components/ui/list/index.ts
@@ -1,1 +1,0 @@
-export { List } from './list';
